renamed portfolioSliceTypes to portfolioSliceCommon
